Add files via upload
The file poll84.csv contains responses from a US exit poll that took place before the 1984 US elections. It contains the following variables:
    ◦ vote to take the values 0 if the respondent expresses the intention to vote for Democrats, 1 for Republicans.
    ◦ income, a variable taking the values lowest, low, moderate, high, highest.
    ◦ age, the age of the respondent.
    ◦ female, taking the values 0 for males, 1 for a female respondent.
a) Producing a table of summary statistics of the age for each sex of the respondents. 
b) Investigating the voting intention per income level:
    1. by creating a contingency table 
    2. by creating an appropriate plot  
c) Using the appropriate hypothesis test, deciding if the voting intention of the poll is compatible with the election result (approximately 58% voted for Republicans, 40% for Democrats). Stating the hypothesis for the test/tests and offering my interpretation. 
d) Investigating the distribution of the age per income level
e) Using ANOVA to investigate if the average age is equal among the income levels. 
(where appropriate, assuming a significance level of 5%)